refactor: use .graphql extension for post mutations
