+++ src/apollo/posts/mutations/DeletePost.mutation.graphql
@@ -1,0 +1,3 @@
+mutation DeletePost($id: Int!) {
+  deletePost(id: $id)
+}
