+++ src/apollo/posts/mutations/UpdatePost.mutation.graphql
@@ -1,0 +1,15 @@
+mutation UpdatePost($postData: UpdatePostInput!) {
+  updatePost(postData: $postData) {
+    post {
+      id
+      body
+      user {
+        ...UserAvatar
+      }
+      group {
+        ...GroupAvatar
+      }
+      createdAt
+    }
+  }
+}
