--- src/apollo/posts/mutations/UpdatePost.mutation.ts
@@ -1,22 +1,0 @@
-import { gql } from "@apollo/client";
-import { GroupAvatarFragmentDoc, UserAvatarFragmentDoc } from "../../gen";
-
-gql`
-  mutation UpdatePost($postData: UpdatePostInput!) {
-    updatePost(postData: $postData) {
-      post {
-        id
-        body
-        user {
-          ...UserAvatar
-        }
-        group {
-          ...GroupAvatar
-        }
-        createdAt
-      }
-    }
-  }
-  ${GroupAvatarFragmentDoc}
-  ${UserAvatarFragmentDoc}
-`;
