+++ src/components/Posts/PostCard.tsx
@@ -18,7 +18,6 @@ import {
   useDeletePostMutation,
   useMeQuery,
 } from "../../apollo/gen";
-import { removePost } from "../../apollo/posts/mutations/DeletePost.mutation";
 import {
   MIDDOT_WITH_SPACES,
   NavigationPaths,
@@ -34,6 +33,7 @@ import AttachedImageList from "../Images/AttachedImageList";
 import ItemMenu from "../Shared/ItemMenu";
 import Link from "../Shared/Link";
 import UserAvatar from "../Users/UserAvatar";
+import { removePost } from "./DeletePostButton";
 import PostCardFooter from "./PostCardFooter";
 
 const CardHeader = styled(MuiCardHeader)(() => ({
